// frontend/src/shared/api/generated/model/testTemplateOut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * novalearn-api
 * OpenAPI spec version: 0.1.0
 */
import type { TestTemplateOutGenerationProgress } from './testTemplateOutGenerationProgress';
import type { TestTemplateOutQuestionTypeCounts } from './testTemplateOutQuestionTypeCounts';

/**
 * Template list-view.
 */
export interface TestTemplateOut {
  id: string;
  user_id: string | null;
  folder_id: string;
  lesson_id?: string | null;
  name: string;
  type: string;
  status: string;
  node_ids?: string[] | null;
  total_questions: number;
  total_marks: number | null;
  mark_scheme?: string | null;
  original_filename?: string | null;
  created_at: string;
  generation_progress?: TestTemplateOutGenerationProgress;
  question_type_counts?: TestTemplateOutQuestionTypeCounts;
}